Day_01/slackbot/requirements.txt
slacker
beautifulsoup4
websocket_client
requests
